Make uHEAP_ALIGNMENT uint-typed so RoundUp deduces on x64 (x64 fix)
uHEAP_ALIGNMENT was #define'd as sizeof(int), whose type is size_t. The
CFastHeap ctor and pvAllocate call RoundUp(uint_size, uHEAP_ALIGNMENT);
RoundUp is template<class T> T RoundUp(T, T), so it needs both args the
same type. On Win32 size_t == uint so it deduced; on x64 size_t is 64-bit
!= uint, so deduction failed (C2672), cascading to 312 x64 errors across
every TU that instantiates a fast heap.

Cast the macro to uint ((uint)sizeof(int)) so both RoundUp args are uint.
Value is unchanged (4); its other uses are all modulo checks, fine as
uint. No-op on Win32 (green). On the diagnostic x64 build the FastHeap
RoundUp errors clear and the count drops 1347 -> 1037.

Next x64 root cause: more custom operator new(uint) sites needing size_t
(C2821 x218: Instance.hpp, FrameHeap.hpp, ...), and the renderer-primitive
asm (WalkEx.hpp, FastBumpEx.hpp).

+++ jp2_pc/Source/Lib/Sys/FastHeap.hpp
@@ -554,7 +554,7 @@ template<class T> inline void* operator new(size_t u_size, CDArray<T>& ma)
 
 
 // The default heap alignment used.
-#define uHEAP_ALIGNMENT	sizeof(int)
+#define uHEAP_ALIGNMENT	((uint)sizeof(int))	// (uint)-typed so RoundUp(uint_size, uHEAP_ALIGNMENT) deduces on x64, where sizeof is size_t != uint
 
 //*********************************************************************************************
 //
